jpa 적용, test 코드 수정

+++ src/main/java/antigravity/controller/ProductController.java
@@ -27,10 +27,10 @@ public class ProductController {
     }
 
     private ProductInfoRequest getParam() {
-        int[] couponIds = {1, 2};
+        long[] couponIds = {1L, 2L};
 
         ProductInfoRequest request = ProductInfoRequest.builder()
-                .productId(1)
+                .productId(1L)
                 .couponIds(couponIds)
                 .build();
